demo using the OSS tier
TK: limits and feature usage reporting

Doing literally the most throwaway scaffolding of a "sidecar"
imaginable.  (It's not a sidecar.  It's not even a bicycle.  It's
child_process.spawn.)

Technically we could get limits now, but the demo doesn't actually need
that, it needs limit *checking*, which can't really be done until we
have usage reporting.

Payment method is done using the traditional Stripe.elements flow,
pulling the CustomerID from Tier via the metadata when it's created by
tier.subscribe.

So, pretty slow/janky, but we're actually using it for most stuff now.